06-lab-complete/requirements.txt
fastapi==0.115.0
uvicorn[standard]
pydantic==2.9.0
pyjwt==2.9.0
python-dotenv==1.0.1
redis==5.1.0
psutil==6.0.0
